requirements.txt
Django>=5.2,<6.0
psycopg2>=2.9.9
gunicorn>=21.2.0
whitenoise>=6.0.0
djangorestframework>=3.15.0
